+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added:
 
-- DRIFT-528: statistic information about time series, [PR-6](https://github.com/panda-official/DriftProtocol/pull/6)  
+- DRIFT-528: Statistic information about time series, [PR-6](https://github.com/panda-official/DriftProtocol/pull/6)
+- DRIFT-551: Documentation on RTD, [PR-7](https://github.com/panda-official/DriftProtocol/pull/7)
 
 ## 0.2.0 - 2022-08-23